Move on-chain crypto into chain crate
All things specified in babbage/crypto.cddl are now back in the chain
crate.

To work around some things e.g. signatures, hashes being uses in the
crypto crate we have those being their base locations with chain/crypto
having a `ChainCrypto<T>` wrapper around this.

We have the to/from raw bytes in a `RawBytesEncoding` trait to help out
as well for when things are raw bytes.

All crypto bindings are still in the `wasm` crate. We will need to look
into a separate `crypto-wasm` crate and first test out how those would
interact and if it's possible/a good idea.
